Switched openssl to netcat to test ports.
Previously openssl failed on handshake because of missing client
certificates, which failed the whole test. This can be fixed by using
netcat to test open ports.

It is important to use "</dev/null" as stdin as "-z" test flag in netcat
doesn't exist in the latest version of netcat anymore.